hidden input field and session message
if the update button has been clicked, update the record with new $name and $location using the value from the hidden id input field. Set value of $id to 0 outside of if statement.

Set the SESSION message to updated. Set the message type to redirect you back to index.php